Change the "Enable All" button to "Reset Defaults"
This is probably what the user wants, since some of the rules that are off by
default are quite buggy.  The "Enable All" button is actually still lurking
there, but the XUL doesn't seem to support an extra3 button.  Does XUL have
some arbitrary limit on the number of buttons it supports?