Abandon the "git archive" XPI production method.
It was hard to exclude things while using it, and seemed impossible to add the
default.rulesets agglomerated XML file that we produce at build time but which
is not in the repo.

Also, the semantics of "make" vs "make uncommitted" was continuing to cause
unsuspecting users misery on a regular basis.