Issue #4060:
Added toolbar button for global enable/disable of the addon. It works by simply removing and re-adding the listeners for now. No graphical feedback as to what the status is.

TO DO:  1) Graphic feedback showing addon status {enabled, disabled} 2) Verify there are no left-over policy bits anywhere that need to be removed.