Revert "Pedantic tweaks to the header text in Chrome's popup."
In the spirity of pedantry :) ...

"Applied" -> "Applicable" : this is a list of rules that may or may not be applied to this
page/site.  They might all be off!

"webpage" -> "site" : rulesets may be listed even though they didn't actually
affect this specific pageload, and would only affect other pages on the site.

This reverts commit d7b82a89a3cdd26b4470477b5ae4aabc5b1a8a8e.