Remove www.linuxfoundation.org rules, add more LF lists.*
www.linuxfoundation.org has to-HTTP redirects all across it, with redirects to HTTPS for certain pages and any users with sessions (complete with a distince logged-in session cookie with secure cookie): having https-everywhere play well with that is unlikely.

Match both lists.linux-foundation.org and lists.linuxfoundation.org but redirect them to the version without the hyphen.

Add lists subdomain for SPDX and FOSSBazaar projects.