14 of 16 rulesets submitted by tlseverywhere.
https://mail1.eff.org/pipermail/https-everywhere-rules/2012-April/001108.html
2 of the rulesets: eu1.snoobi.com and speech.leseweb.dk do not seem to serve any
kind of content. eu1.snoobi.com may have been added because it seems to serve
cookies on several of the websites included here and it appears to be an
analytics service.

I'm still not sure what the leseweb rule is for. Hopefully tlseverywhere can get
back to me about this.