Update GUI and manifest for Chrome 20 and above.
The page popup has been modified to match the look-and-feel of Chrome's
latest Settings page. Additionally, the manifest.json has been updated
to the version 2 format. Version 1 was deprecated in Chrome 18. CSS
and Javascript had to be moved out of popup.html and into separate
files in order to comply with version 2.

Tested on:
* Debian Linux, Chrome 20
* Mac OS X Lion, Chrome 21
* Windows 7, Chrome 21