[OpenTTD] remove binaries rules
From IRC:

<Rubidium> hi, the ruleset for openttd.org (especially .binaries.openttd.org)
is and has been broken. You redirect the URLs our mirror balancer gives out
back into the mirror balancer. This might have worked in the past when the
amount of redirects triggered it to go to another mirror, but now it mostly
fails with 404 errors.

<Rubidium> there is one mirror that is not "affected" since we may not call
that <xyz>.binaries.openttd.org, but for all other mirrors it simply fails