[Boston Globe] boston.com default off for cert mismatch
They now have Akamai hosting without a valid cert.  This is new
in the last day or two, perhaps as a result of high traffic levels
and changed site configuration in the aftermath of the Boston
Marathon bombing.  We should check to see if the HTTPS support
returns in the future.