Be more careful about finding HSTS duplicate rules
chromium-preloads.py needs to check whether a rule in the hsts preload
list applies to a target already in the https-e rules, and, if so,
ignores the hsts rule. Previously this was done by matching the hsts
rule name (ex: google.com) anywhere in the content of the https-e
xml rule files. This commit changes the script to find exact matches between
hsts rule names and https-e targets, properly handling wildcards.

Before this change, chromium-preloads.py created 62 new rules. After,
it created 48.