[Doubleclick.net] Enable
Assuming that an appreciable number of HTTPSE users don't block
doubleclick.net, enabling this ruleset is hugely beneficial given the
domain's prevalence.

The commit disabling this ruleset gives no details or links regarding
its rationale. Presumably, not every resource that exists on
*.doubleclick.net causes the problem cited, and as there are no details
available--yet--limiting coverage to paths that don't cause problems
might be somewhat difficult.

Furthermore, based on how quickly google seems to address the problems
mentioned in our rulesets, it wouldn't be surprising if the problem has
been fixed over the past two years.

With that, enabling in the hope that:

	- Any further reports will give more detail; and
	- The problem will have been fixed by now anyway