Fix for issue #9511 (https://trac.torproject.org/projects/tor/ticket/9511)
The toolbar_button_binding.xml file already defines the class values (toolbarbutton-1 and chromeclass-toolbar-additional) for the toolbar button. No need for the toolbar_button.xul
file to also define these values. My guess is that these classes (toolbarbutton-1 and chromeclass-toolbar-additional) add extra padding because they are in an XBL binding.