Optimize onBeforeSendHeaders
This function makes little sense to me. I think the logic originally was:

If we lose the race to update cookies in onHeadersReceived, we can win it now in
onBeforeSendHeaders.

Unless I'm missing something, I think this *whole function* might be completely useless.

I'll open up an issue to discuss that in a bit.

Signed-off-by: Nick Semenkovich <semenko@alum.mit.edu>