Move displayPageAction out of the hot onBeforeRequest
displayPageAction is called (unnecessarily) potentially /hundreds/ of times per tab.

It only needs to be called /once/ when a page updates and has applicable rules.

We'd be fine just removing it from onBeforeRequest (and depending on chrome.tabs.onUpdated) except for the super-edge case where a tab:
- Meets zero potentially applicable rulesets
- Has absolutely no rulesets that apply to dependencies (included JS, etc.)
- Makes a request we /can/ apply rules to, but it happens after DOM document complete when tabs.onUpdated has already fired

In practice, I was unable to find *any* sites that met this criteria, but I can make a localhost test case.

This adds a 10 sec. timer to check if active tabs are missing the pageAction icon.

Signed-off-by: Nick Semenkovich <semenko@alum.mit.edu>