Fire pageAction only from tabs.onUpdated/onReplaced
pageAction is responsible for adding the HTTPS Everywhere icon next to the URL bar. We used to display this only on pages w/ applicable rules (almost every page), but that logic was messy and slow. Now we display the icon all the time.

This might get called once or twice per-tab, but it's relatively low cost. At least we're no longer calling it multiple times for every request :)

Signed-off-by: Nick Semenkovich <semenko@alum.mit.edu>