Allow users to create user-specified rules. Rules are saved in local storage and automatically loaded when the browser / extension starts up.
TODO:
This change does not allow users to edit/delete existing rules or to submit them to a central location.
This change doesn't support exceptions or cookie rules yet (but no reason they couldn't be easily added).

I also introduced a structure for a JS object that can be used to represent rules.