Alexa ruleset checker working
The ruleset checker seems to be working: it downloads and unzips the
Alexa Top1M and automatically generates the git diff.
Comparing the two seems to be working as well.

Manually checked some rules: they seem to have been correctly identified
as in the Top 1M and not in stable