make new updateUrl for Android (temporarily)
If we point to the old updateUrl, then people who accidentally install
desktop stable HTTPS Everywhere on their phones will succeed and then
discover that HTTPS Everywhere doesn't work. This is because addons
check the updateUrl for compatibility information if install.rdf returns
negative.

Eventually, we want to release a single xpi that is both compatible with
Android and desktop Firefox. Hopefully this will happen in 3.4.6, after
which the Android-specific updateUrl is obsolete.