[makexpi.sh] Make python interpreter explicit again
bde023b3 removed explicit interpreter invocation, perhaps because the
author thereof thought it redundant.

However, trusted path execution (TPE), wherein execution of paths that
are writeable by any role other than the superuser is denied, is thereby
broken. TPE is very useful for making sure that untrusted interpreters
are forbidden *even when* MAC is disabled (e.g., when the host is
updating policies, or initialization fails).

	- Making it explicit doesn't break anything (AFAICS)
	- I received no replies about this on IRC; and
	- No relevant changes have been committed

So this is my naïve fix.