Expand NewsBlur rule to cover blurblogs
NewsBlur provides what it calls "Blurblogs", which are separate per-user
subdomains displaying each user's shared stories in a blog-like format. HTTPS
is available on these subdomains, as well as the already known www and dev
subdomains. However, there is also an 'icons' subdomain which uses an invalid
(Amazon S3) certificate.