Add rule for GitHub Pages.
GitHub Pages now supports HTTPS. I believe this is recent, but I might simply have not noticed it before. The changes add a rule to the existing ruleset for GitHub and adds `github.io` to the targets of the ruleset. The rule added has been tested shallowly but it is unlikely that it causes problems since `github.io` is used exclusively for GitHub Pages, which only hosts static websites.