Remove exclusion of dumps.wikimedia.org
dumps.wikimedia.org is currently excluded from the Wikimedia ruleset, but it supports SSL. There might be a reason I am not aware of to exclude it, but it does not appear to be so.

Some of the other subdomains excluded redirect to dumps.wikimedia.org. It might be worth it to redirect them directly to dumps.wikimedia.org and enable SSL at the same time to avoid a request being sent without SSL to the subdomain first.