Use LC_ALL=C for filename sorting in make-sqlite.
Sorting is necessary for deterministic builds. Per
https://trac.torproject.org/projects/tor/ticket/11630#comment:19,
gk's build machine doesn't have the UTF-8 locale. Since the locale
doesn't determine output encoding, only sorting, and all that matters
is that the sort order is deterministic, LC_ALL=C is fine, and more
likely to be present on any given machine.