Don't rewrite from https
Unless it fixes problems that we cause.

Peter was right -- this practice was more of a problem than it's worth.

From now on, when needed to fix problems that our rewrites create, be
sure to make to make it blindingly obvious with a comment above each
instance.

This will probably:

* Fix undiscovered loops
* Prevent confusion over what 'https?' means and its utility