Improve detection of updated rulesets.
Previous, we tried to detect when rulesets changed by using git log --since.
However, this doesn't work very well for uncommited changes, and it works poorly
on Travis, which does a shallow clone and therefore is not able to get a good
history of which files have changed since when.

Instead, we create a list of SHA256 hashes of files as they stood on 2014-02-12
(plus some extras that slipped through because the changes were made before then
but merged later). When running coverage tests, and files that match those
hashes are ignored.

Also add a test-generator script to aid in generating test URLs for rulesets.