Revert all Google rules to pre-checker version.
There were two major bugs in Google rules related to recent changes:

  - https://github.com/EFForg/https-everywhere/issues/1216
  - https://github.com/EFForg/https-everywhere/issues/1215

The simplest fix is to simply revert these rules to the way they were before
I disabled them and @StevenRoddis rewrote them to comply with new testing and
style guides.

We should roll this change forward again after the 5.0 release and fix the
specific bugs, but this is a quick fix so we can proceed with the release.