Amazon.com: Fix exclusion
Firstly, per @jsha in ce1b13b2eaa78f93025c90e5fd2fafba4980c532, '?!'
does not mean 'not'. Rather, it means 'must not', and regardless is not
a binary relation to the greediness of '+'.

For whatever reason, way back when, I started using it as though it were.

(thanks @jsha for pointing out my error)

Secondly, there's an erroneous empty case, so the exclusion had none of
the desired effect.