Remove www.php.net from the list of SSL-enabled domains.
GeoDNS has been enabled on the www.php.net subdomain, which means that
requests are now going to be routed to any one of php.net's many
geographically distributed mirrors. As we don't intend to distribute our
private key to our large number of mirror providers, HTTPS requests will
either be denied or result in a certificate mismatch error.