Remove Enable / Disable rules window.
Loading all the rulesets at once for the Enable / Disable rules window
causes browser lockup and sometimes crashes. All of the functionality
in the window is now available in other places: Rules can be enabled
and disabled directly from the drop down menu. "Disable all" is
available as the menu item "Disable HTTPS Everywhere." And viewing
the source of rulesets is available through the HTTPS Everywhere Atlas.

The one missing piece of functionality was "Reset rulesets to defaults," which I
moved to a menu item.

Now, the preferences link in Firefox's Add-On pane will take you to the SSL
Observatory preferences.