Version 5.1.0
With AMO signature.

Uses a constant git commit for make-sqlite.

This is required so that when we build the sqlite DB, it exactly matches the
version submitted to AMO for signing. In future versions, we'll just omit the
git commit table.

Also, this change includes a spurious .install.rdf.swp that was accidentally
included in the version submitted for AMO signing. Rather than wait two days
to get re-signed I've just included it.