Change id of extension for AMO signing.
To get an automated signature from AMO for self-hosting, we need to have
a separate unlisted version of our extension. Because of a problem in AMO,
the unlisted version cannot have the same id as the listed version. So we're
renaming the self-hosted version to https-everywhere-eff@eff.org.

This also includes a change to make-sqlite.py to remove the git commit table.
The new release process will require an initial build, then a signature, then a
second build. Since that second build will be at a different git commit,
rulesets.sqlite would change and cause the signature to no longer match.

Conflicts:
	utils/make-sqlite.py