[Avm.de] Add exclusions
This excludes most of the content on avm.de, adding the existing exclusions from en.avm.de to this main domain as well. This fixes a redirect loop on the front page and false mixed content blocking on several subpages (cf. the test URLs). In particular, this fixes https://github.com/EFForg/https-everywhere/issues/3346.