Implement async reading of rules from SQLite.
This borrows a hack from AdBlock Plus: since the replaceChannel APIs expect a
synchronous response, any request for which we don't yet know whether we want to
rewrite it gets redirected to its own URL, and channel gets suspended. Once we've
loaded the necessary state from SQLite and can decide on a rewrite based on
in-memory information only, we unsuspend the channel. The internal redirect
fires and we get another opportunity to rewrite that request.