Add a few strings for localisation in the 'about' window
This changes a few hardcoded strings so as to allow for localisation. Obviously the way this is done is not ideal, though, as translations are handled through a torproject.org based git submodule. The corresponding changes to this, which in turn is based on Transifex in a way that I don't remember, would be of the same form as those in https://github.com/fuglede/torproject_translation/commit/67875f8bf227798d8b46b77a4f6557e70cd365c3 which contain the changes to the English and Danish language strings.