Partially revert #3883.
This caused a regression: Rules that are active but not taking effect (formerly
moot rules) are simply not listed, when my intention was to list them (i.e.,
match the Chrome behavior).

Will roll forward again when I have a fix, but didn't want to leave things in a
broken state until then.

This is a partial revert because the Chrome changes were fine, so I'm leaving
them.