Increase timeout limits
This doubles the length of the timeout restrictions for all current https-everywhere-checker configs in an attempt to fix https://github.com/EFForg/https-everywhere/issues/4810.

Arguably, since Travis is the one causing that particular issue, it could suffice to change only http.checker.config, but increasing all of them is probably safer for database-wide runs anyway.