[Custhelp.com] Update targets
This redirects directly to the target that the host itself redirects to through a middle-man, when accessed via http. Normally, we don't carry out the redirects for the sites, if we can instead upgrade to a secure connection directly (as the ruleset did previously), but in this case, the total amount of redirects were causing issues for the fetch tests (see https://github.com/EFForg/https-everywhere/pull/4747#issuecomment-233090118), and since the initial redirect already bounced to a different domain, I figured this wasn't as bad an idea in this case, as it might be in others.