[Archive.is] Update ruleset (#5886)
* [Archive.is] Update ruleset

After having had a couple of different issues with this ruleset, cf. in particular

    https://github.com/EFForg/https-everywhere/issues/5805 and
    https://github.com/EFForg/https-everywhere/issues/4075

I prompted the webmaster for what would be a proper solution moving forwards. On their request, this avoids redirecting anything to https://archive.is/ but ensures that the relevant connections are still upgrading by using archive.fo instead.

* [Archive.is] Update comment