[Energinet.dk] Fix various issues
The existing ruleset caused a redirect loop due to the inclusion on the following script on all content containing pages;

	<script type="text/javascript">
	{
	  var url=location.href;
	  if(url.indexOf("https://") > -1)
	  {
	    var newurl=url.replace("https://","http://");
	    location.href=newurl;
	  }
	}
	</script>

This fixes that by exclusion of those pages.

Moreover, clicking the "English"/"Danish" link, you'd end up on a page for editing the webpage contents, which is great in itself.