Updated fronter.com rules to match recent developments. Every subdomain under fronter.com now supports (and redirects to) HTTPS, so I've removed some conditionals, and replaced them with a wildcard entry. At the same time I've added some explicit test URLs to ensure coverage of what's behind the wildcard, and also removed conditionals for securecookie.
The root of the bottom two test URLs only return 403, but I left
them in in case you wanted >3. If testing against a 403 is no good
(the redirect still works fine, though), feel free to get rid of
them.

I've tried following the style guide, and hope the result is to
your satisfaction.