Exclude easyjet.com/ejcms/cache15m/api/.*
I get the following error in Chrome when this is rewritten:
XMLHttpRequest cannot load
https://www.easyjet.com/ejcms/cache15m/api/flights/search?AdminFeePerBookin…,LGW,LTN,SEN,STN&PreferredOriginIatas=LGW,LTN,SEN,STN&StartDate=2017-02-01.
Credentials flag is 'true', but the 'Access-Control-Allow-Credentials'
header is ''. It must be 'true' to allow credentials. Origin
'http://www.easyjet.com' is therefore not allowed access.

This breaks the functionality of the website.

Also, expand wildcard, because travis is complaining at me.