Restore some W3C rules
Commit 9bf49cbeb4696cdea76df23f685da1d11031f693 disabled the entire ruleset,
apparently on the basis that test.w3.org no longer exists on the server. I
agree that test.w3.org no longer exists, but to disable the entire ruleset
because of just one of its subdomains seems to be throwing the baby out with
the bathwater.

Most of the other subdomains still exist and have HTTPS available, so I
re-enabled them.

I deleted the test.w3.org subdomain because it no longer exists.

I left two subdomains disabled because I am unsure how to handle
them:
- the "people" subdomain, which serves different content depending on
  whether it is accessed via HTTP or HTTPS.
- the "validator-suite" subdomain, which redirects to a page hosted by
  GitHub.