Move some handler tests to the parse module test
These were end-to-end tests, but really what we were asserting on is
that the parse module could handle those cases. In the end-to-end
tests all we care about is that the handlers have the correct behavior
given a successful result from the parse module, regardless of how
much the parser had to clean up the input. So move these tests were
they belong, which is _just_ passing stuff directly to the parser
module, instead of calling through everything in reality.