Improve trivialize-rules to detect compex rules
Explode regular expressions and see if it results in a small list of static URLs
that can be trivialized.

Note: rulesets with <securecookie /> are skipped for now because trivializing
those can lead to certain hosts being removed from targets even though cookies
should still be secured there. In future, we can analyze securecookie as well
to detect static matches and trivialize more rulesets.