Assert on API calls, not context creation
Before the Sinon spy for `context.issue` didn't return anything. That
meant that in the source code, things like
`context.github.issue.createComment()` would get `undefined` passed to
it (since that was the return value of `context.issue`).

Because e.g. `createComment` would be passed `undefined`, we couldn't
actually make any assertions on it, so we asserted on the context
creation instead. This works okay except that it means that different
Perjury topics would interfere with each other - any topic that was
run later would have to account for the earlier calls to issue
creation.

The obvious solution to this is to make the mock behavior match the
real implementation more closely, so that's what we
do. `context.issue` now just passes through the values it was given,
which in practice means that things like `createComment` are basically
directly passed whatever was given to `context.issue()`. With that
change, we're now able to actually assert something that makes sense -
e.g. if we're trying to assert that no comments were posted, we can
now actually do that instead of faking it and asserting that
`context.issue` was only called once (instead of not at all, because
of some previous action like a label change or something).

As a bonus, this also means that we can remove some magic numbers
which frankly I have _never_ understood!