Fix settings on new versions of Chrome.
On Chrome Canary (currently version 67.0.3368.0), the checkboxes in the
HTTPS Everywhere popup are missing and this error shows up in the
background page's console every time the popup loads:

    Error in event handler: TypeError: Error processing argument at index -1, conversion failure from undefined

It looks like this is because `chrome.storage`'s methods no longer work
unbound. The fix is to bind them to the `StorageArea ` when they get
pulled out.