Updated URL to new Atlas URL
Chose to use https://atlas.eff.org/index.html instead of https://atlas.eff.org/ because links to go back to "HTTPS Everywhere Atlas" main page from the Atlas itself go to the index.html file.  This is also a little safer in the event of a server misconfiguration or hack.