Add tekram.com
Rule only targets the two domains that don't use redirect.
In the backend, tekram seems to use bigcommerce for everything,
which supports https and in the site it's being included as https.

It's strange that they set the hsts header with max-age=0, yet
backend consistently use https for everything.

Added all cookies as securecookies, while there are cookies
with the flag, some SESSION ones weren't.